- fix error in userblacklist DB scheme definition, the whitelist colum type   should be of TINYINT (or equivalent) size - no DB version update necessary, as the module in question is not released yet
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@4184 689a6050-402a-0410-94f2-e92a70836424

+++ tools/postgres/userblacklist-create.sql
@@ -4,7 +4,7 @@ CREATE TABLE userblacklist (
     username VARCHAR(64) DEFAULT '' NOT NULL,
     domain VARCHAR(64) DEFAULT '' NOT NULL,
     prefix VARCHAR(64) DEFAULT '' NOT NULL,
-    whitelist INTEGER DEFAULT 0 NOT NULL,
+    whitelist SMALLINT DEFAULT 0 NOT NULL,
     description VARCHAR(64) DEFAULT '' NOT NULL
 );
 
@@ -14,7 +14,7 @@ INSERT INTO version (table_name, table_version) values ('globalblacklist','1');
 CREATE TABLE globalblacklist (
     id SERIAL PRIMARY KEY NOT NULL,
     prefix VARCHAR(64) DEFAULT '' NOT NULL,
-    whitelist INTEGER DEFAULT 0 NOT NULL,
+    whitelist SMALLINT DEFAULT 0 NOT NULL,
     description VARCHAR(64) DEFAULT '' NOT NULL
 );
